+++ src/algo/FindLowSimilar.hpp
@@ -9,6 +9,7 @@
 #define NPGE_FIND_LOW_SIMILAR_HPP_
 
 #include "BlocksJobs.hpp"
+#include "Decimal.hpp"
 
 namespace npge {
 
@@ -33,7 +34,7 @@ public:
     typedef std::vector<Region> Regions;
 
     /** Return weight factor corresponding to min_identity */
-    static int get_weight_factor(double min_identity);
+    static int get_weight_factor(Decimal min_identity);
 
     /** Create list of regions from list of color.
     Weight factor is applied to bad regions.
